Add Heuristics module.
