+++ plutus-core/plutus-ir/src/PlutusIR/Compiler.hs
@@ -44,7 +44,7 @@ import PlutusIR.Compiler.Types
 import PlutusIR.Error
 import PlutusIR.Transform.Beta qualified as Beta
 import PlutusIR.Transform.DeadCode qualified as DeadCode
-import PlutusIR.Transform.Inline qualified as Inline
+import PlutusIR.Transform.Inline.Inline qualified as Inline
 import PlutusIR.Transform.LetFloat qualified as LetFloat
 import PlutusIR.Transform.LetMerge qualified as LetMerge
 import PlutusIR.Transform.NonStrict qualified as NonStrict
