+++ plutus-core/plutus-ir/test/TransformSpec.hs
@@ -20,7 +20,7 @@ import PlutusIR.Parser
 import PlutusIR.Test
 import PlutusIR.Transform.Beta qualified as Beta
 import PlutusIR.Transform.DeadCode qualified as DeadCode
-import PlutusIR.Transform.Inline qualified as Inline
+import PlutusIR.Transform.Inline.Inline qualified as Inline
 import PlutusIR.Transform.LetFloat qualified as LetFloat
 import PlutusIR.Transform.LetMerge qualified as LetMerge
 import PlutusIR.Transform.NonStrict qualified as NonStrict
